MNT: move server utils into top of engine.py
